requirements.txt
fastapi
uvicorn
numpy
opencv-python
google-cloud-vision
pytest
pytest-mock
httpx
python-multipart
celery
redis
flower